+++ src/pwm.rs
@@ -266,7 +266,10 @@ macro_rules! hal {
                 }
 
                 fn get_duty(&self) -> u16 {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr1.read().ccr().bits() }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr1.read().ccr1().bits() }
                 }
 
                 fn get_max_duty(&self) -> u16 {
@@ -274,7 +277,10 @@ macro_rules! hal {
                 }
 
                 fn set_duty(&mut self, duty: u16) {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr1.write(|w| w.ccr().bits(duty)) }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr1.write(|w| w.ccr1().bits(duty)) }
                 }
             }
 
@@ -290,7 +296,10 @@ macro_rules! hal {
                 }
 
                 fn get_duty(&self) -> u16 {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr2.read().ccr().bits() }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr2.read().ccr2().bits() }
                 }
 
                 fn get_max_duty(&self) -> u16 {
@@ -298,7 +307,10 @@ macro_rules! hal {
                 }
 
                 fn set_duty(&mut self, duty: u16) {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr2.write(|w| w.ccr().bits(duty)) }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr2.write(|w| w.ccr2().bits(duty)) }
                 }
             }
 
@@ -314,7 +326,10 @@ macro_rules! hal {
                 }
 
                 fn get_duty(&self) -> u16 {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr3.read().ccr().bits() }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr3.read().ccr3().bits() }
                 }
 
                 fn get_max_duty(&self) -> u16 {
@@ -322,7 +337,10 @@ macro_rules! hal {
                 }
 
                 fn set_duty(&mut self, duty: u16) {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr3.write(|w| w.ccr().bits(duty)) }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr3.write(|w| w.ccr3().bits(duty)) }
                 }
             }
 
@@ -338,7 +356,10 @@ macro_rules! hal {
                 }
 
                 fn get_duty(&self) -> u16 {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr4.read().ccr().bits() }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr4.read().ccr4().bits() }
                 }
 
                 fn get_max_duty(&self) -> u16 {
@@ -346,7 +367,10 @@ macro_rules! hal {
                 }
 
                 fn set_duty(&mut self, duty: u16) {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr4.write(|w| w.ccr().bits(duty)) }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr4.write(|w| w.ccr4().bits(duty)) }
                 }
             }
         )+